+++ Plugins/AM_MetaboliteDetector_Plugin/AnalysisToolRunnerMetaboliteDetector.cs
@@ -19,8 +19,8 @@ namespace AnalysisManagerMetaboliteDetectorPlugin
     /// </summary>
     public class AnalysisToolRunnerMetaboliteDetector : AnalysisToolRunnerBase
     {
-        private const float PROGRESS_PCT_STARTING = 5;
-        private const float PROGRESS_PCT_COMPLETE = 99;
+        private const int PROGRESS_PCT_STARTING = 5;
+        private const int PROGRESS_PCT_COMPLETE = 99;
 
         private const string METABOLITE_DETECTOR_CONSOLE_OUTPUT = "MetaboliteDetector_ConsoleOutput.txt";
